Add a script for running salt
This adds a salt_utils module which can run salt on the master,
gather and parse the results to decide if state execution succeeded.
Otherwise running a salt state always produces a 0 error code.

I've modified the bootstrap script to clone this repo so that the
new module is available to be called from fabric during deploy.