Make getting connection objects easier in fab_tasks
We were returning more and more things from get_config, not all of which was used everywhere (and we wanted to add 2 more) and that was just getting silly. Most of what we were returning was only used to pass in to EC2 or Cloudformation constructors, so lets just make it easier to create those instances.