Enforce PEP8 on any changed hunks.
An enforcable stlye guide is a good thing - we have some developers who
are less experienced with python so being able to say 'do it this way'
is a good thing.

I am open to disabling some more rules (for instance I have set line
lengths to 160 up from the default 79) but not the enforcement of a
style.

This change means that we don't have to fix PEP8 style violations in one
big go, and can still get useful automated style checking.