Generate Userdata script to set hostname of instances based on template
i.e. If we set hostname_pattern to "{instance_id}.my_project" then we
drop a script in via userdata that [cloud-init](http://cloudinit.readthedocs.org/)
will instruct later phases of cloud-init to set the hostname. So that
sudo can resolve the hostname and and we don't get warnings/delays we
also want to be able to instruct cloud_config to manage /etc/hosts -
which we do via letting the user pass in a chunk of YAML for
[cloud_config](http://cloudinit.readthedocs.org/en/latest/topics/examples.html)

There are other types of userdata we could add to the Userdata "Mime
Multipart archive" but we don't need them right now so don't expose them
from the config YAML