Check for AWS_ROLE_ARN_ID in utils:connect_to_aws before assuming role
connect_to_aws checks for the cross-account profile and then tries to
assume a role based on that. This means that if are not using the
cross-account as a profile name we will skip assuming roles. This change
will also attempt to assume a role if we have a AWS_ROLE_ARN_ID variable
set.