Added elasticache with engine redis
Can now add an elasticache section to the template to get a redis
based elasticache. The number of clusters in the replication group,
node_type and port can be specified. A list of redis backup files
can be provided also, to seed the new elasticache resource.
The replication groups name and engine is provided as an output.

elasticache:
  clusters: 3
  node_type: cache.m1.small
  port: 10024
  seeds:
    s3:
    - test-redis-bucket-947923urhiuy8923d/redis.rdb