Add enter and exit maintenance fab tasks
Adds two fab tasks called enter_maintenance which takes a maintenance_ip as
a parameter and exit_maintenance. This is slightly complicated as there is
nothing stopping us from having multiple internet facing ELBs, so we kind
of have to loop through our stack and config to find the right ones.
When exiting maintenance in order to get the dns_name we need to get the
real load balancer object from AWS so we need to work our way to that by:

a) create a mapping of logical name into physical name for our stack's elbs
b) obtain all ELB object, filtering for physical names that belong to our
   stack
and then for each elb:

1) taking the config name from the cloudformation yaml
2) converting it to the logical name using get_elb_safe_name()
   (effectively converts something like shiny-foo into ELBshinyfoo)
3) traverse (b) to lookup the dns_name for the logical name