Make update_certs task use set_ssl_certificates retry logic
The update certs task calls time.sleep(3) to overcome the problem
with checking something that AWS believes the ssl cert to be available,
but it takes a few seconds to be 'eventually consistent'.
The ELB::set_ssl_certificates has retry logic to handle this case
but it is currently unused by the task. This change makes update_certs
retry 3 times so that AWS has consistent knowledge on the state of certs
and the action will succeed.

(Closes #196)