Add option to specify additional ELB aliases
This change adds the option of having additional entries in the
cloudformation config specifying ELB aliases. For example

elb:
  - name: myelb
    aliases:
      - thiselb
      - yourelb

The above setup would create route53 A records as aliases to the
elb name address. If the aliases already exist, they will not
be overwritten by default, although there is the option to forcibly
do so.