Rename S3 bucket output to be consistent
When we create cloudformation things we can create Outputs that represent them,
which then get turned into grains, meaning that we can do lookups on randomly
named resources. The buckets that are created have 2 things, a name, and a policy.
The output for the S3 bucket currently is named like it references the policy,
but actually it references the name. This change makes it more coherent with the
form '<bucket_id>BucketName'.