Currently when deleting tail blocks forever
bootstrap_cfn 's utils.tail() blocks forever on stack deletion.
Added a helper function to detect when the status is deleted or
failed to delete (to represent the cfn_delete event finishing).
Note that after delete_done it may be possible that the stack
still has lingering resources, that's why we leave it vague,
and only suggest that the 'cfn delete event' finished.