Allow multi-part policies on aditional s3 buckets
AIM policy documents are made up of a list of policy statements.
Currently bootstrap-cfn only allows a single custom policy for an S3
bucket. This change allows a list of custom policies to be provided.
Backwards compatibility is maintained by allowing either a single policy
dict or a list of policies as an array.

I have only implemented this on the additional s3 buckets because the
static buckets are for a specific use case and should be kept as simple
as possible.